Components are not being sent to the Client
Currently The player shows as they have the NBT data for the item, but they do not render with the wings on.
In here is shows that the entity has the old "TrinketComponent" wings, and they also get replace with the new wings.
But placing a debugger in Icarus "WingsFeatureRenderer", The "TrinketComponent" does not show that the entity has the wings.